requirements.txt
matplotlib==3.5.3
numpy==1.22.4
opencv_python==4.6.0.66
Pillow==9.4.0
scikit_image==0.19.3
scipy==1.10.1
tensorflow==2.11.0
tf_slim==1.1.0
typer